Refactor interface structure: move and rename screen and request interfaces for better organization
