+++ src/interfaces/requests/save-score-request.ts
