+++ src/interfaces/requests/verify-registration-request.ts
