+++ src/interfaces/responses/authentication-response.ts
