+++ src/interfaces/responses/ranking-response.ts
