+++ src/interfaces/responses/registration-options-response.ts
