+++ src/interfaces/screens/game-screen.ts
