+++ src/interfaces/screens/screen-manager.ts
