+++ src/screens/main-screen.ts
@@ -1,7 +1,7 @@
 import { BaseGameScreen } from "./base/base-game-screen.js";
 import { ScreenManagerService } from "../services/screen-manager-service.js";
 import { MainBackgroundObject } from "../objects/backgrounds/main-background-object.js";
-import type { GameScreen } from "../interfaces/screen/game-screen.js";
+import type { GameScreen } from "../interfaces/screena/game-screen.js";
 import type { GameState } from "../models/game-state.js";
 
 export class MainScreen extends BaseGameScreen {
