+++ src/services/screen-manager-service.ts
@@ -1,5 +1,5 @@
-import type { GameScreen } from "../interfaces/screen/game-screen.js";
-import type { ScreenManager } from "../interfaces/screen/screen-manager.js";
+import type { GameScreen } from "../interfaces/screena/game-screen.js";
+import type { ScreenManager } from "../interfaces/screena/screen-manager.js";
 import { ScreenTransitionService } from "./screen-transition-service.js";
 
 export class ScreenManagerService implements ScreenManager {
