+++ src/services/screen-transition-service.ts
@@ -1,5 +1,5 @@
-import type { GameScreen } from "../interfaces/screen/game-screen.js";
-import type { ScreenManager } from "../interfaces/screen/screen-manager.js";
+import type { GameScreen } from "../interfaces/screena/game-screen.js";
+import type { ScreenManager } from "../interfaces/screena/screen-manager.js";
 
 export class ScreenTransitionService {
   private elapsedTransitionMilliseconds: number = 0;
